Initial commit of the PerfAggregator class.
PerfAggregator is a class that, given connections to a list of vCenter
Servers, will fetch the entire VM folder and ResourcePool hierarchies,
including all VIM::VirtualMachine objects and aggregate VM stats along
the tree hierarchies. The PerfAggregator class allows for users to
perform post processing on the data returned by vCenter, e.g. to augment
it with addtional data that was obtained using a combination of
VM annotations (or custom values) and an external DB. Post processing
can also define additional tree structures that may be completely
independent of the VM folder and ResourcePool hirarchies provided by
vCenter, e.g. one based on VMs used for testing of a set of source code
branches.