Make tests work in 1.8
This involves several hacks to get it working. In 1.8.7, 'test-unit'
is included in the Ruby distribution. It is not included and is a
separate Gem in 1.9+ (though test-unit's gemspec does not note this).
Also, I had to remove a single test from running in 1.8.7 due to it
failing because of ordering. I made a note to come back to this.