Fix documentation formatting
I believe YARD parsed the documentation correctly, but I've never really
liked the formatting here. In the near future, I plan to go back and fix
it throughout the project, but I just accepted a PR on this method and
needed to add the @options for the new :no_delta option and thought I
might as well fix the rest while I'm there.