Remove the overriden `network` on `VirtualMachine`
This method was proposed and accepted in #96. The other code in this
are not overriden methods, but additional methods that are not included
in the API. This was accepted mistakenly believing that this code was
adding a new method, not overriding an already available method.

This commit removes the override and calls the API's method. If the
functionality added in #96 is desired, it should be done by the caller
or added as a new method that does not override the API's method.

Fixes #108