Correct vmodl.db changes
I fixed an issue (#127) by updating the vmodl.db and it appears in
fixing that issue, I created another one (the previous issue was just
reopened).

In updating the vmodl.db for the original issue, it duplicated some info
that was already there and caused an issue. This update removes the
unnecessary duplication and fixes #127 again.

Signed-off-by: J.R. Garcia <jrg@vmware.com>