Duplicate 'sharing' attribute from VirtualDiskFlatVer2BackingInfo
The vmodl.db from current master (462126e), appears to have a duplicate 'sharing' attribute on the VirtualDiskFlatVer2BackingInfo object. This breaks cloning virtual machines through the vSphere API in both vsphere-fog and rbvmomi when running against vSphere 6.5 (and potentially other revisions). This patch removes this duplicate definition as it does not appear in the API WSDL.