Allow deployOvf to function even when DNS is not fully configured.
We observed an issue when hosts were added to vCenter via a FQDN, but
DNS was not functioning end to end in the environment, specifically at
the location where the deployOvf call was initiated.  Thus deploying an
Ovf failed.  This updates the upload command to always ensure a valid
location by IP even when DNS is not available at the calling location:w.