Prepare for new release
Updating version and contributors list for new release. Unfortunately,
this involved a change to a dependency that has been renamed (with the
former being deprecated). We have a file that makes modifications for
this dependency and this is a breaking change due to the changes we
expose for that dependency. Because of this, we have had to bump the
major version to 2.0.0 (rather than just 1.13.1 or 1.14.0). Doing this
will prevent any breaking changes for people following semantic
versioning in their Gemfile/gemspec.

Signed-off-by: J.R. Garcia <jrg@vmware.com>