Only ship the necessary libs in the gem artifact
The gem artifact is not the ideal place to ship examples, contributing
docs, a rake file and various hidden files. This will greatly reduce the
size of the installed gem artifact and speed up installs on systems with
malware scanners. This drops the compressed gem size from 221k to 47k.

Signed-off-by: Tim Smith <tsmith@chef.io>