Don't install bundler in travis before_install (#152)
An appropriate bundler version is installed by rvm with the ruby version
requested.  Later if we gem install bundler it will fail on really old
rubies (< 2.3) because it tries to install bundler 2.0.

As long as the specific version of bundler doesn't matter we can just
use the bundler version that rvm installs.